Remove dependency on local redspot

+++ scripts/start-substrate.sh
@@ -42,22 +42,22 @@ if [ -z "$SUBSTRATE_CONTAINER_NAME" ]; then
 fi
 
 SUBSTRATE_PID=$(docker exec $SUBSTRATE_CONTAINER_NAME sh -c "ps aux | grep -v grep | grep substrate-contracts-node | awk '{print \$2}'")
-if [[ $TEST_DB == true ]]; then
-  echo "Killing existing substrate..."
-  docker exec "$SUBSTRATE_CONTAINER_NAME" kill -9 "$SUBSTRATE_PID"
-fi
 
 # restart chain with no data
 if [[ $RESTART_CHAIN == true ]]; then
   if [[ -n "$SUBSTRATE_PID" ]]; then
     echo "Killing existing substrate..."
     docker exec "$SUBSTRATE_CONTAINER_NAME" kill -9 "$SUBSTRATE_PID"
-  fi;
+  fi
   docker exec -d "$SUBSTRATE_CONTAINER_NAME" rm -rf ./chain-data
-fi;
+fi
 
 # switch db
 if [[ $TEST_DB == true ]]; then
+  if [[ -n "$SUBSTRATE_PID" ]]; then
+    echo "Killing existing substrate..."
+    docker exec "$SUBSTRATE_CONTAINER_NAME" kill -9 "$SUBSTRATE_PID"
+  fi
   docker exec -d "$SUBSTRATE_CONTAINER_NAME" rm -rf ./chain-test
   docker cp ./.chain-test/. "$SUBSTRATE_CONTAINER_NAME":/chain-test
   docker exec -d "$SUBSTRATE_CONTAINER_NAME" substrate-contracts-node --dev -d ./chain-test --unsafe-ws-external --rpc-external --prometheus-external -linfo,runtime::contracts=debug
